requirements.txt
requests
click 
pytest   
pyyaml
prettytables
ja4
scapy